Add comments to regex.sh
This commit added comments to the line of code used in creating r1_output.txt to give a better understanding as to how it works. Additionally, this commit changes the wording in the comments left for the line concerning r0_output.txt to show sample output in a clearer manner.